Clarify the working of with-packages
It took me a really long time (almost a day) to figure out how to use this, being unfamiliar with the way idris uses packages and with limited experience with nix expressions. The usage turned out to be rather simple but it would help if it's spelled out more. Also, if I actually include prelude and base in the desired libraries I get an error so I think that remark can be removed.